Added a seperate file for Fona Utilities
Pulled exisiting FONA files into their own .c and .h file
Added Utilites to:
*Read messages sent to the FONA
*Interpret received message.  Right now it only understands messages
telling the system to update the RTCC clock time
* Clear received message buffer on the SIM

In IWP Utilities included an interrupt to read from the UART
