+++ nbproject/configurations.xml
@@ -19,6 +19,10 @@
       <itemPath>I2C.h</itemPath>
       <itemPath>Pin_Manager.c</itemPath>
       <itemPath>Pin_Manager.h</itemPath>
+      <itemPath>FONAUtilities.h</itemPath>
+      <itemPath>FONAUtilities.c</itemPath>
+      <itemPath>UnusedFunctions.c</itemPath>
+      <itemPath>UnusedFunctions.h</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
